+++ packages/provider/test/selector_test.dart
@@ -77,33 +77,195 @@ void main() {
       throwsAssertionError,
     );
   });
+  testWidgets('Deep compare maps by default', (tester) async {
+    var value = <int, int>{};
+    final builder = MockBuilder<Map<int, int>>();
+    when(builder(any, any, any)).thenReturn(Container());
+
+    final selector = Selector0<Map<int, int>>(
+      selector: (_) => value,
+      builder: builder,
+    );
+    await tester.pumpWidget(selector);
+
+    verify(builder(argThat(isNotNull), value, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    final element = tester.element(find.byWidget(selector));
+
+    value = {0: 0, 1: 1};
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verify(builder(argThat(isNotNull), value, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    value = {0: 0, 1: 1};
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verifyNoMoreInteractions(builder);
+  });
+  testWidgets('Deep compare iterables by default', (tester) async {
+    var value = <int>[].whereType<int>();
+    final builder = MockBuilder<Iterable<int>>();
+    when(builder(any, any, any)).thenReturn(Container());
+
+    final selector = Selector0<Iterable<int>>(
+      selector: (_) => value,
+      builder: builder,
+    );
+    await tester.pumpWidget(selector);
+
+    verify(builder(argThat(isNotNull), value, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    final element = tester.element(find.byWidget(selector));
+
+    value = [1, 2].whereType<int>();
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verify(builder(argThat(isNotNull), value, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    value = [1, 2].whereType<int>();
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verifyNoMoreInteractions(builder);
+  });
+  testWidgets('Deep compare sets by default', (tester) async {
+    var value = <int>{};
+    final builder = MockBuilder<Set<int>>();
+    when(builder(any, any, any)).thenReturn(Container());
+
+    final selector = Selector0<Set<int>>(
+      selector: (_) => value,
+      builder: builder,
+    );
+    await tester.pumpWidget(selector);
+
+    verify(builder(argThat(isNotNull), value, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    final element = tester.element(find.byWidget(selector));
+
+    value = {1, 2};
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verify(builder(argThat(isNotNull), value, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    value = {1, 2};
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verifyNoMoreInteractions(builder);
+  });
+  testWidgets('Deep compare lists by default', (tester) async {
+    var value = <int>[];
+    final builder = MockBuilder<List<int>>();
+    when(builder(any, any, any)).thenReturn(Container());
+
+    final selector = Selector0<List<int>>(
+      selector: (_) => value,
+      builder: builder,
+    );
+    await tester.pumpWidget(selector);
+
+    verify(builder(argThat(isNotNull), value, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    final element = tester.element(find.byWidget(selector));
+
+    value = [1, 2];
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verify(builder(argThat(isNotNull), value, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    value = [1, 2];
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verifyNoMoreInteractions(builder);
+  });
+  testWidgets('custom shouldRebuid', (tester) async {
+    var value = 0;
+    var shouldRebuild = true;
+    final mockShouldRebuild = MockShouldRebuild<int>();
+    when(mockShouldRebuild(any, any)).thenAnswer((_) => shouldRebuild);
+
+    final builder = MockBuilder<int>();
+    when(builder(any, any, any)).thenReturn(Container());
+
+    final selector = Selector0<int>(
+      selector: (_) => value,
+      shouldRebuild: mockShouldRebuild,
+      builder: builder,
+    );
+    await tester.pumpWidget(selector);
+
+    verifyZeroInteractions(mockShouldRebuild);
+    verify(builder(argThat(isNotNull), 0, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    final element = tester.element(find.byWidget(selector));
+
+    value = 1;
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verify(mockShouldRebuild(0, 1)).called(1);
+    verifyNoMoreInteractions(mockShouldRebuild);
+    verify(builder(argThat(isNotNull), 1, null)).called(1);
+    verifyNoMoreInteractions(builder);
+
+    shouldRebuild = false;
+    value = 2;
+    element.markNeedsBuild();
+    await tester.pump();
+
+    verify(mockShouldRebuild(1, 2)).called(1);
+    verifyNoMoreInteractions(mockShouldRebuild);
+    verifyNoMoreInteractions(builder);
+  });
   testWidgets('passes `child` and `key`', (tester) async {
     final key = GlobalKey();
-    await tester.pumpWidget(Selector0<Null>(
-      key: key,
-      selector: (_) => null,
-      builder: (_, __, child) => child,
-      child: const Text('42', textDirection: TextDirection.ltr),
-    ));
+    await tester.pumpWidget(
+      Selector0<Null>(
+        key: key,
+        selector: (_) => null,
+        builder: (_, __, child) => child,
+        child: const Text('42', textDirection: TextDirection.ltr),
+      ),
+    );
 
     expect(key.currentContext, isNotNull);
 
     expect(find.text('42'), findsOneWidget);
   });
   testWidgets('calls builder if the callback changes', (tester) async {
-    await tester.pumpWidget(Selector0<int>(
-      selector: (_) => 42,
-      builder: (_, __, ___) =>
-          const Text('foo', textDirection: TextDirection.ltr),
-    ));
+    await tester.pumpWidget(
+      Selector0<int>(
+        selector: (_) => 42,
+        builder: (_, __, ___) =>
+            const Text('foo', textDirection: TextDirection.ltr),
+      ),
+    );
 
     expect(find.text('foo'), findsOneWidget);
 
-    await tester.pumpWidget(Selector0<int>(
-      selector: (_) => 42,
-      builder: (_, __, ___) =>
-          const Text('bar', textDirection: TextDirection.ltr),
-    ));
+    await tester.pumpWidget(
+      Selector0<int>(
+        selector: (_) => 42,
+        builder: (_, __, ___) =>
+            const Text('bar', textDirection: TextDirection.ltr),
+      ),
+    );
 
     expect(find.text('bar'), findsOneWidget);
   });
@@ -113,16 +275,18 @@ void main() {
     var builder = (BuildContext _, int __, Widget child) => child;
     final child = const Text('foo', textDirection: TextDirection.ltr);
 
-    await tester.pumpWidget(MultiProvider(
-      providers: [
-        Selector0<int>(
-          key: key,
-          selector: selector,
-          builder: builder,
-        ),
-      ],
-      child: child,
-    ));
+    await tester.pumpWidget(
+      MultiProvider(
+        providers: [
+          Selector0<int>(
+            key: key,
+            selector: selector,
+            builder: builder,
+          ),
+        ],
+        child: child,
+      ),
+    );
 
     final widget = tester.widget(
       find.byWidgetPredicate((w) => w is Selector0<int>),
@@ -130,8 +294,14 @@ void main() {
 
     expect(find.text('foo'), findsOneWidget);
     expect(widget.key, key);
-    expect(widget.selector, equals(selector));
-    expect(widget.builder, equals(builder));
+    expect(
+      widget.selector,
+      equals(selector),
+    );
+    expect(
+      widget.builder,
+      equals(builder),
+    );
   });
   testWidgets(
       "don't call builder again if it rebuilds"
@@ -142,10 +312,12 @@ void main() {
       return Text(value.toString(), textDirection: TextDirection.ltr);
     });
 
-    await tester.pumpWidget(Selector0<int>(
-      selector: selector,
-      builder: builder,
-    ));
+    await tester.pumpWidget(
+      Selector0<int>(
+        selector: selector,
+        builder: builder,
+      ),
+    );
 
     verify(selector(argThat(isNotNull))).called(1);
     verifyNoMoreInteractions(selector);
@@ -175,10 +347,12 @@ void main() {
       return Text(value.toString(), textDirection: TextDirection.ltr);
     });
 
-    await tester.pumpWidget(Selector0<int>(
-      selector: selector,
-      builder: builder,
-    ));
+    await tester.pumpWidget(
+      Selector0<int>(
+        selector: selector,
+        builder: builder,
+      ),
+    );
 
     verify(selector(argThat(isNotNull))).called(1);
     verifyNoMoreInteractions(selector);
@@ -337,6 +511,10 @@ class F with _ToString {}
 
 class MockSelector<T> extends Mock {
   T call(BuildContext context);
+}
+
+class MockShouldRebuild<T> extends Mock {
+  bool call(T prev, T next);
 }
 
 class MockBuilder<T> extends Mock {